Drawer Nav connected to map branch
in drawer navigation we no longer need the map constant we can just link it directly to the map screen
